requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fsspec==2025.5.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
openai==1.82.1
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
PyMuPDF==1.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
zipp==3.22.0
